requirements.txt
numpy==1.26.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.3.2
xgboost==2.0.3
catboost==1.2.7
lightgbm==4.5.0
plotly==5.18.0
streamlit==1.37.1
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
joblib==1.3.2
tqdm==4.66.1
requests==2.31.0
pydantic==2.7.3
tabulate==0.9.0
imblearn==0.0